backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pg8000==1.31.2
pydantic-settings==2.7.0
python-dotenv==1.0.1
pdfplumber
PyPDF2
openai==1.58.1
google-generativeai==0.8.3
python-multipart==0.0.20
websockets==12.0
python-docx==1.1.2
